Use the Spring Initializr
I rewrote the document and adjusted the code (mostly changed package names) to use the Spring Initializr.
