+++ complete/src/test/java/com/example/springbootdocker/HelloWorldConfigurationTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hello;
+package com.example.springbootdocker;
 
 import static org.junit.Assert.*;
