+++ initial/src/main/java/com/example/springbootdocker/SpringBootDockerApplication.java
@@ -1,12 +1,13 @@
-package hello;
+package com.example.springbootdocker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class Application {
+public class SpringBootDockerApplication {
+
 	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
+		SpringApplication.run(SpringBootDockerApplication.class, args);
 	}
 
 }
