+++ initial/src/test/java/com/example/springbootdocker/SpringBootDockerApplicationTests.java
@@ -1,0 +1,16 @@
+package com.example.springbootdocker;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SpringBootDockerApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
